predictions_service/requirements.txt
contourpy==1.3.1
cycler==0.12.1
fonttools==4.55.3
grpcio==1.68.1
grpcio-tools==1.68.1
joblib==1.4.2
kiwisolver==1.4.7
matplotlib==3.10.0
numpy==2.2.0
packaging==24.2
pillow==11.0.0
protobuf==5.29.1
psycopg2==2.9.10
pyparsing==3.2.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
scikit-learn==1.6.0
scipy==1.14.1
setuptools==75.6.0
six==1.17.0
threadpoolctl==3.5.0
